# authcm/下/operations/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "身份验证 API",
    "description": "用于用户登录与注册",
    "version": "1.0"
  },
  "produces": ["application/json"],
  "basePath": "/",
  "security": [
    {
      "cmjwt": []
    }
  ],
  "parameters": {
    "apiKey": {
      "name": "cm-api-key",
      "description": "平台发行的 API_KEY",
      "in": "header",
      "type": "string",
      "required": true
    },
...
  },
  "paths": {
...
    "/api/users/{id}": {
      "get": {
        "x-swagger-router-controller": "user",
        "operationId": "findOne",
        "description": "获取用户详细信息",
        "tags": ["resource"],
        "parameters": [
          {
            "$ref": "#/parameters/apiKey"
          },
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "put": {
        "x-swagger-router-controller": "user",
        "operationId": "update",
        "description": "更新用户信息",
        "tags": ["resource"],
        "parameters": [
          {
            "$ref": "#/parameters/apiKey"
          },
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "required": true
          },
          {
            "name": "data",
            "in": "body",
            "description": "用户信息",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      },
      "delete": {
        "x-swagger-router-controller": "user",
        "operationId": "destroy",
        "description": "删除用户",
        "tags": ["resource"],
        "parameters": [
          {
            "$ref": "#/parameters/apiKey"
          },
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "properties": {
        "id": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "cmjwt": {
      "type": "apiKey",
      "name": "cm-api-key",
      "in": "header"
    }
  }
}